feat: better docs split
