+++ umh-core/pkg/communicator/api/v2/http/internal/latency.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package http
+package internal
 
 import (
 	"net"
